# docs/api/openapi.yaml
openapi: 3.1.0
info:
  title: Floppy Disk API
  version: "1.0.0"
  description: |
    REST API for the Floppy Disk cloud storage / sharing / streaming product.

    ## Base URL
    All application endpoints are under `/api/v1/`. In dev the SPA reaches the
    backend same-origin via the Vite proxy.

    ## Authentication
    Phase 1 uses **session auth**: sign in via `POST /api/v1/auth/login`, which
    sets a `sessionid` cookie. Unsafe methods (POST/PUT/PATCH/DELETE) on
    authenticated endpoints also require the CSRF token: call
    `GET /api/v1/auth/csrf` once to obtain the `csrftoken` cookie and echo its
    value in the `X-CSRFToken` header.

    Programmatic clients (e.g. the MCP server) authenticate with an **API key**
    via `Authorization: Bearer <key>` (see the `apiKey` security scheme) - this
    bypasses CSRF.

    ## Conventions
    - IDs are UUIDs unless noted (share tokens are strings).
    - Sizes are bytes.
    - List endpoints return arrays; a few wrap results in an object (noted per path).
    - Errors use `{ "detail": "...", "code": "..." }` (code is present on typed errors).
servers:
  - url: /api/v1
    description: Application API (relative to host)
...
security:
  - sessionAuth: []
    csrfToken: []
  - apiKey: []          # programmatic clients (MCP / integrations) - bypasses CSRF
components:
  securitySchemes:
    sessionAuth:
      type: apiKey
      in: cookie
      name: sessionid
    csrfToken:
      type: apiKey
      in: header
      name: X-CSRFToken
    apiKey:
      type: http
      scheme: bearer
      description: "Programmatic access (MCP / integrations). Send `Authorization: Bearer <key>`."
...
paths:
...
  /public/share/{token}/download:
    get:
      tags: [Sharing]
      summary: Download a shared file's bytes (no account)
      description: >-
        Streams the bytes in local mode (Range-aware) and redirects to a
        presigned URL in R2 mode. For password-protected links, pass the
        password as `?password=` (or the `X-Share-Password` header) - the token
        alone does not bypass the gate.
      security: []
      parameters:
        - { name: token, in: path, required: true, schema: { type: string } }
        - { name: password, in: query, required: false, schema: { type: string } }
      responses:
        "200": { description: File bytes (200 or 206 for a Range request). }
        "302": { description: Redirect to a presigned URL (R2 mode). }
        "401": { description: Incorrect/missing password. }
        "404": { description: Unknown token / nothing to download. }
        "410": { description: Expired or revoked. }

  # ---------------------------------------------------------------- Notifications
...